#6509 Handle tranlation links via the hw.telescope view
The logic for displaying links to translations is currently in a tal
portlet and is rather convoluted and inefficient. I've added the
relevant logic into a view method instead. This view returns the data
structure required to display the links. The portlet needs to be
edited to take advantage of this. The url field in the dictionary
already takes the telescope path (passed in the query string),
into account, which resolves this issue.